Feat: Adds auth route, controllers and queries

+++ backend/src/api/user/user.query.ts
@@ -1,41 +1,76 @@
 import prisma from '../../config/db.js'
-import type { User, UserType } from '@prisma/client'
+import type { User } from '@prisma/client'
+import { UserSchemaType } from '../../validators/auth.schema'
 
-export type CreateUserInput = {
-  type: UserType
-  isCompany: boolean
-  firstName?: string
-  lastName?: string
-  companyName?: string
-  taxNumber?: string
-  email: string
-  phone: string
-  addressId?: number | null
-}
+export const createUser = async (data: UserSchemaType): Promise<User> => {
+  const { address, ...restData } = data
 
-export type UpdateUserInput = Partial<CreateUserInput>
+  let addressId: number | null = null
 
-export const createUser = async (data: CreateUserInput): Promise<User> => {
-  return prisma.user.create({ data })
-}
+  if (address) {
+    const createdAddress = await prisma.address.create({ data: address })
+    addressId = createdAddress.id
+  }
 
-export const getUsers = async (): Promise<User[]> => {
-  return prisma.user.findMany()
+  return prisma.user.create({
+    data: {
+      ...restData,
+      addressId: addressId
+    }
+  })
 }
 
 export const getUserById = async (id: number): Promise<User | null> => {
   return prisma.user.findUnique({ where: { id } })
 }
 
-export const updateUser = async (id: number, data: UpdateUserInput): Promise<User> => {
+export const getUserByEmail = async (email: string): Promise<User | null> => {
+  return prisma.user.findUnique({ where: { email } })
+}
+
+export const updateUser = async (
+  id: number,
+  data: Partial<UserSchemaType & { addressId: number | null }>
+): Promise<User> => {
+  const { address, ...resetData } = data
+
+  if (address) {
+    if (resetData.addressId) {
+      await prisma.address.update({
+        where: { id: resetData.addressId },
+        data: address
+      })
+    } else {
+      const createdAddress = await prisma.address.create({ data: address })
+      resetData.addressId = createdAddress.id
+    }
+  }
+
   return prisma.user.update({
     where: { id },
-    data,
+    data: resetData
+  })
+}
+
+export const changePassword = async (id: number, hashedPassword: string): Promise<User> => {
+  return prisma.user.update({
+    where: { id },
+    data: { password: hashedPassword }
   })
 }
 
 export const deleteUser = async (id: number): Promise<User> => {
+  const user = await prisma.user.findUnique({ where: { id } })
+  if (!user) {
+    throw new Error(`User with id ${id} not found`)
+  }
+  if (user.addressId) {
+    // If the user has an address, delete it
+    await prisma.address.delete({
+      where: { id: user.addressId }
+    })
+  }
   return prisma.user.delete({
-    where: { id },
+    where: { id }
   })
 }
